requirements.txt
# Core
flask==3.1.0
flask-cors==5.0.0
gunicorn==23.0.0
python-dotenv==1.0.1
pydantic>=2.9.0
requests==2.32.3

# API Documentation
flasgger==0.9.7.1
flask-swagger-ui==4.11.1

# AI Provider
google-generativeai==0.8.3
openai==1.12.0

# Security
flask-limiter==3.8.0
pyjwt==2.9.0
cryptography==43.0.3

# Monitoring
prometheus-client==0.21.0
structlog==24.4.0
redis==5.2.0

# Database (without psycopg2 for Windows, will be in Docker)
sqlalchemy==2.0.35

# Testing
pytest==8.3.3
pytest-mock==3.14.0
pytest-cov==5.0.0

# Retry logic
tenacity==9.0.0

# Utilities
bleach==6.1.0
validators==0.22.0

# Email validation for Pydantic
email-validator==2.1.1
